Extract NodeRemoval feature from GraphProjector
