+++ Neos.ContentGraph.DoctrineDbalAdapter/Classes/Domain/Projection/GraphProjector.php
@@ -15,9 +15,9 @@ namespace Neos\ContentGraph\DoctrineDbalAdapter\Domain\Projection;
 
 use Doctrine\DBAL\Connection;
 use Neos\Cache\Frontend\VariableFrontend;
+use Neos\ContentGraph\DoctrineDbalAdapter\Domain\Projection\Feature\NodeRemoval;
 use Neos\ContentGraph\DoctrineDbalAdapter\Domain\Repository\ProjectionContentGraph;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\CopyableAcrossContentStreamsInterface;
-use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Event\NodeAggregateWasRemoved;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Event;
 use Neos\EventSourcedContentRepository\Domain as ContentRepository;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Event\NodePropertiesWereSet;
@@ -51,6 +51,8 @@ use Neos\Flow\Annotations as Flow;
  */
 class GraphProjector implements ProjectorInterface, AfterInvokeInterface
 {
+    use NodeRemoval;
+
     const RELATION_DEFAULT_OFFSET = 128;
 
     /**
@@ -1065,59 +1067,6 @@ insert ignore into neos_contentgraph_restrictionrelation
                 }
             }
         });
-    }
-
-    /**
-     * @param NodeAggregateWasRemoved $event
-     * @throws \Throwable
-     */
-    public function whenNodeAggregateWasRemoved(NodeAggregateWasRemoved $event)
-    {
-        // the focus here is to be correct; that's why the method is not overly performant (for now at least). We might
-        // lateron find tricks to improve performance
-        $this->transactional(function () use ($event) {
-            $this->removeOutgoingRestrictionRelationsOfNodeAggregateInDimensionSpacePoints(
-                $event->getContentStreamIdentifier(),
-                $event->getNodeAggregateIdentifier(),
-                $event->getAffectedCoveredDimensionSpacePoints()
-            );
-
-            $ingoingRelations = $this->projectionContentGraph->findIngoingHierarchyRelationsForNodeAggregate(
-                $event->getContentStreamIdentifier(),
-                $event->getNodeAggregateIdentifier(),
-                $event->getAffectedCoveredDimensionSpacePoints()
-            );
-            foreach ($ingoingRelations as $ingoingRelation) {
-                $this->removeRelationRecursivelyFromDatabaseIncludingNonReferencedNodes($ingoingRelation);
-            }
-        });
-    }
-
-    /**
-     * @param HierarchyRelation $ingoingRelation
-     * @throws \Doctrine\DBAL\DBALException
-     */
-    protected function removeRelationRecursivelyFromDatabaseIncludingNonReferencedNodes(HierarchyRelation $ingoingRelation)
-    {
-        $ingoingRelation->removeFromDatabase($this->getDatabaseConnection());
-
-        foreach ($this->projectionContentGraph->findOutgoingHierarchyRelationsForNode($ingoingRelation->childNodeAnchor, $ingoingRelation->contentStreamIdentifier, new DimensionSpacePointSet([$ingoingRelation->dimensionSpacePoint])) as $outgoingRelation) {
-            $this->removeRelationRecursivelyFromDatabaseIncludingNonReferencedNodes($outgoingRelation);
-        }
-
-        // remove node itself if it does not have any incoming edges anymore
-        $this->getDatabaseConnection()->executeUpdate('
-            DELETE n FROM neos_contentgraph_node n
-                LEFT JOIN
-                    neos_contentgraph_hierarchyrelation h ON h.childnodeanchor = n.relationanchorpoint
-                WHERE
-                    n.relationanchorpoint = :anchorPointForNode
-                    AND h.contentstreamidentifier IS NULL
-                ',
-            [
-                'anchorPointForNode' => (string)$ingoingRelation->childNodeAnchor,
-            ]
-        );
     }
 
     /**
